[HOTSPOT-281] refactor: 폴더 네임 수정
